website/serializers.py
from django.db.models import fields
from rest_framework import serializers
from .models import *

class BlogSerializer(serializers.ModelSerializer):
    class Meta:
        model = Blog
        fields = '__all__'


class CareerCVSerializer(serializers.ModelSerializer):
    class Meta:
        model = CareerCV
        fields = '__all__'


class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = '__all__'
