requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.9
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
djangorestframework==3.12.4
future==0.18.2
html5lib==1.1
idna==3.3
Pillow==8.4.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
